+++ build-logic/convention/src/main/kotlin/corp/tbm/cleanwizard/buildLogic/convention/foundation/extensions/DependencyHandlerScopeExtensions.kt
