+++ foundation/annotations/src/main/kotlin/corp/tbm/cleanwizard/foundation/annotations/Enum.kt
