+++ foundation/codegen/kotlinpoet/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/kotlinpoet/Constants.kt
