+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/EnumType.kt
@@ -1,6 +1,7 @@
 package corp.tbm.cleanarchitecturemapper.foundation.codegen.universal
 
 import corp.tbm.cleanarchitecturemapper.foundation.annotations.*
+import corp.tbm.cleanwizard.foundation.annotations.*
 import kotlin.reflect.KClass
 
 enum class EnumType(
