+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/exceptions/references/PropertyAlreadyMarkedWithEnum.kt
