+++ foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/extensions/ksp/ks/KSClassDeclarationExtensions.kt
